+++ CHANGELOG.md
@@ -10,6 +10,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 ### Removed
 
+## [1.3.0] - 2019-03-05
+### Added
+- Add ability that returns relays status
+- Add 3 ways connection schema to documentation
+- Add domapic-relays-switch bin command and script
+
 ## [1.2.0] - 2019-01-06
 ### Changed
 - Upgrade domapic-service version. Connection will not work with domapic-controller versions lower than 1.0.0-alpha.14.